Revert "Restructured Component Tree to Lazy Load"
