+++ client/src/component-tree/App.js
@@ -1,15 +1,15 @@
-import React, { lazy, Suspense } from "react";
+import React from "react";
 import "../styling/index.scss";
 import { useDarkMode } from "../hooks/useDarkMode";
 import { Switch, Route } from "react-router-dom";
 
 //Component Imports
 import Loader from "./Loader/Loader.js";
-// import Nav from "./Nav/Nav.js";
-// import Home from "./Home/Home.js";
-// Lazy Components
-const Nav = lazy(() => import("./Nav/Nav.js"));
-const Home = lazy(() => import("./Home/Home.js"));
+import Nav from "./Nav/Nav.js";
+import About from "./About/About.js";
+import Home from "./Home/Home.js";
+import Projects from "./Projects/Projects.js";
+import Dashboard from "./Tutorials/Dashboard.js";
 
 function App() {
   //Initialize dark mode hook
@@ -22,15 +22,36 @@ function App() {
 
   return (
     <div className="app-container is-dark">
-      <Suspense fallback={Loader}>
-        <Switch>
-          <Route exact path="/">
-            <Loader />
-            <Nav darkMode={darkMode} setDarkMode={setDarkMode} />
-            <Home darkMode={darkMode} />
-          </Route>
-        </Switch>
-      </Suspense>
+      <Switch>
+        <Route exact path="/">
+          <Loader />
+          <Nav darkMode={darkMode} setDarkMode={setDarkMode} />
+          <Home darkMode={darkMode} />
+        </Route>
+
+        <Route path="/about">
+          <Loader />
+          <Nav darkMode={darkMode} setDarkMode={setDarkMode} />
+          <About darkMode={darkMode} />
+        </Route>
+
+        <Route path="/projects">
+          <Loader />
+          <Nav darkMode={darkMode} setDarkMode={setDarkMode} />
+          <Projects darkMode={darkMode} />
+        </Route>
+
+        <Route path="/home">
+          <Loader />
+          <Nav darkMode={darkMode} setDarkMode={setDarkMode} />
+        </Route>
+
+        <Route path="/tutorials">
+          <Loader />
+          <Nav darkMode={darkMode} setDarkMode={setDarkMode} />
+          <Dashboard darkMode={darkMode} />
+        </Route>
+      </Switch>
     </div>
   );
 }
